transfer viewBids component to row component
